--- .envconf.js
@@ -1,3 +1,0 @@
-process.env.SDC_URL="https://us-east-1.api.joyentcloud.com";
-process.env.SDC_KEY_ID="/wyatt/keys/d4:19:cc:38:44:a8:5a:aa:76:1c:65:66:ba:08:1e:cb";
-process.env.SDC_KEY_PATH="/Users/wyatt.lyonpreul/.ssh/id_rsa";
