--- packages/cloudapi-gql/lib/api/account.js
@@ -1,12 +1,0 @@
-'use strict';
-
-const CloudApi = require('./cloudapi');
-
-
-exports.get = (args, request) => {
-  return CloudApi('getAccount', args, request);
-};
-
-exports.update = (args, request) => {
-  return CloudApi('updateAccount', args, request);
-};
