--- packages/cloudapi-gql/lib/api/config.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const CloudApi = require('./cloudapi');
-
-
-module.exports = (args, request) => {
-  const options = {
-    path: '/my/config'
-  };
-  return CloudApi('fetch', options, request);
-};
