--- packages/cloudapi-gql/lib/api/errors.json
@@ -1,16 +1,0 @@
-{
-  "BadRequest": "You sent bad HTTP",
-  "InternalError": "Something went wrong in Triton",
-  "InUseError": "The object is in use and cannot be operated on",
-  "InvalidArgument": "You sent bad arguments or a bad value for an argument",
-  "InvalidCredentials": "Authentication failed",
-  "InvalidHeader": "You sent a bad HTTP header",
-  "InvalidVersion": "You sent a bad Api-Version string",
-  "MissingParameter": "You didn't send a required parameter",
-  "NotAuthorized": "You don't have access to the requested resource",
-  "RequestThrottled": "You were throttled",
-  "RequestTooLarge": "You sent too much request data",
-  "RequestMoved": "HTTP Redirect",
-  "ResourceNotFound": "What you asked for wasn't found",
-  "UnknownError": "Something completely unexpected happened!"
-}
