--- packages/cloudapi-gql/lib/api/nics.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const CloudApi = require('./cloudapi');
-
-
-module.exports.list = () => { return request('listNics'); };
-module.exports.get = (ctx) => { return request('getNic', ctx); };
-module.exports.add = (ctx) => { return request(''); };
-module.exports.destroy = (ctx) => { return request(''); };
