--- packages/cloudapi-gql/lib/api/packages.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const CloudApi = require('./cloudapi');
-
-
-module.exports.list = (ctx) => { return request('listPackages', ctx); };
-module.exports.get = ({ id, name }) => { return request.fetch(`/:login/packages/${id || name}`); };
