--- packages/cloudapi-gql/lib/api/services.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const CloudApi = require('./cloudapi');
-
-module.exports = () => { return request('listServices'); };
