--- packages/cloudapi-gql/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "cloudapi-gql",
-  "version": "2.0.0",
-  "license": "MPL-2.0",
-  "repository": "github:yldio/joyent-portal",
-  "main": "lib",
-  "scripts": {
-    "lint": "belly-button --fix",
-    "test": "echo 0",
-    "test-ci": "echo 0",
-    "start": "PORT=4000 node dev-server.js",
-    "dev": "CORS=1 PORT=4000 nodemon dev-server.js",
-    "fake": "graphql-faker ./lib/schema/schema.graphql -p 4000",
-    "prepublish": "echo 0",
-    "graphdoc": "graphdoc -s ./lib/schema/schema.graphql -o ./doc --force",
-    "faker": "graphql-faker ./lib/schema/schema.graphql"
-  },
-  "dependencies": {
-    "apollo-errors": "^1.5.1",
-    "boom": "^6.0.0",
-    "bunyan": "^1.8.12",
-    "dotenv": "^4.0.0",
-    "force-array": "^3.1.0",
-    "hasha": "^3.0.0",
-    "node-fetch": "^1.7.3",
-    "triton": "^5.4.0",
-    "wreck": "^14.0.0"
-  },
-  "devDependencies": {
-    "@2fd/graphdoc": "^2.4.0",
-    "belly-button": "^4.0.1",
-    "execa": "^0.8.0",
-    "graphql-faker": "^1.4.0",
-    "graphql-playground": "^1.0.8",
-    "graphql-tools": "^2.3.0",
-    "graphql-voyager": "^1.0.0-rc.9",
-    "hapi": "^16.6.2",
-    "inert": "^4.2.1"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      "doc/*"
-    ]
-  }
-}
