+++ packages/my-joy-beta/package.json
@@ -6,7 +6,7 @@
   "repository": "github:yldio/joyent-portal",
   "main": "build/",
   "scripts": {
-    "dev": "REACT_APP_GQL_PORT=4000 PORT=3069 REACT_APP_GQL_PROTOCOL=http joyent-react-scripts start",
+    "dev": "REACT_APP_GQL_URL=http://localhost:4000/graphql PORT=3069 joyent-react-scripts start",
     "start": "PORT=3069 joyent-react-scripts start",
     "build": "NODE_ENV=production joyent-react-scripts build",
     "lint-ci":
@@ -19,8 +19,20 @@
   },
   "dependencies": {
     "@manaflair/redux-batch": "^0.1.0",
-    "apollo": "^0.2.2",
-    "joyent-ui-toolkit": "^2.0.1",
+    "eslint": "^4.5.0",
+    "eslint-config-joyent-portal": "^3.0.0",
+    "eslint-config-prettier": "^2.3.0",
+    "eslint-config-react-app": "^2.0.0",
+    "eslint-config-xo-space": "^0.16.0",
+    "eslint-gh-status-reporter": "^1.0.7",
+    "eslint-plugin-flowtype": "^2.35.1",
+    "eslint-plugin-graphql": "^1.3.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-markdown": "^1.0.0-beta.6",
+    "eslint-plugin-prettier": "^2.2.0",
+    "eslint-plugin-react": "^7.3.0",
+    "joyent-ui-toolkit": "^2.0.0",
     "lodash.find": "^4.6.0",
     "lodash.get": "^4.4.2",
     "lodash.isstring": "^4.0.1",
