+++ server.js
@@ -1,19 +1,16 @@
 'use strict';
 
+const CloudApi = require('cloudapi-gql');
+const Crumb = require('crumb');
 const Hapi = require('hapi');
 const HapiPino = require('hapi-pino');
 const Inert = require('inert');
 const Path = require('path');
 const Sso = require('sso-auth');
-const CloudApi = require('./packages/cloudapi-gql');
+const MyJoy = require('./packages/my-joy-beta');
 
 
-const server = new Hapi.Server({
-  debug: {
-    log: ['error'],
-    request: ['error']
-  }
-});
+const server = new Hapi.Server();
 
 const handlerError = (err) => {
   if (err) {
@@ -23,16 +20,8 @@ const handlerError = (err) => {
   }
 };
 
-
 server.connection({
-  port: 80,
-  routes: {
-    cors: {
-      origin: ['localhost', 'localhost:80'],
-      credentials: true,
-      additionalHeaders: ['Cookie']
-    }
-  }
+  port: 80
 });
 
 server.register(
@@ -41,10 +30,10 @@ server.register(
       register: Sso,
       options: {
         cookie: {
-          password: process.env.COOKIE_PASSWORD || 'uUELnoNBVl-Mrey-gTY9uFkQ_0DO3DMP',
+          password: process.env.COOKIE_PASSWORD,
           isSecure: false,
           domain: 'localhost',
-          isHttpOnly: false,
+          isHttpOnly: true,
           ttl: 1000 * 60 * 60       // 1 hour
         },
         sso: {
@@ -54,67 +43,30 @@ server.register(
       }
     },
     {
+      register: Crumb,
+      options: {
+        restful: true,
+        cookieOptions: {
+          isSecure: false,
+          domain: 'localhost',
+          isHttpOnly: false,
+          ttl: 1000 * 60 * 60       // 1 hour
+        }
+      }
+    },
+    {
       register: HapiPino,
       options: {
         prettyPrint: true,
-        logEvents: ['request-error']
+        logEvents: ['request-error', 'error', 'graqhql-error']
       }
     },
     Inert,
-    CloudApi
+    CloudApi,
+    MyJoy
   ],
   (err) => {
     handlerError(err);
-
-    server.route([
-      {
-        method: 'GET',
-        path: '/static/{path*}',
-        config: {
-          auth: false,
-          handler: {
-            directory: {
-              path: Path.join(__dirname, './packages/my-joy-beta/build/static/'),
-              redirectToSlash: true,
-              index: false
-            }
-          }
-        }
-      },
-      {
-        method: 'GET',
-        path: '/{path}.json',
-        config: {
-          auth: false,
-          handler: (request, reply) => {
-            const path = Path.join(__dirname, `./packages/my-joy-beta/build/${request.params.path}.json`);
-            reply.file(path);
-          }
-        }
-      },
-      {
-        method: 'GET',
-        path: '/service-worker.js',
-        config: {
-          auth: false,
-          handler: (request, reply) => {
-            const path = Path.join(__dirname, `./packages/my-joy-beta/build/service-worker.js`);
-            reply.file(path);
-          }
-        }
-      },
-      {
-        method: '*',
-        path: '/{path*}',
-        config: {
-          handler: {
-            file: {
-              path: Path.join(__dirname, './packages/my-joy-beta/build/index.html')
-            }
-          }
-        }
-      }
-    ]);
 
     server.auth.default('sso');
